Refactor: firebase 로그인 로그아웃 함수 파일 분리
